Revert "added lazy router loading"
