+++ src/App.js
@@ -1,30 +1,13 @@
-import React, { Suspense } from 'react';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { Container } from 'reactstrap';
+/* eslint-disable react/prop-types */
+import React from 'react';
 import './App.css';
-import Login from './components/Login/Login';
-import Error from './components/Error/Error';
-import PrivateRoute from './routes/PrivateRoute';
+import { Container } from '@mui/material';
 
-const Departments = React.lazy(() => import('./components/Departments/Departments'));
-const Department = React.lazy(() => import('./components/Department/Department'));
-
-function App() {
+function App({ children }) {
   return (
-    <BrowserRouter>
-      <Container className="container">
-        <Switch>
-          <Route exact path="/" component={Login} />
-          <Suspense fallback={<p>Loading...</p>}>
-            <Switch>
-              <PrivateRoute path="/departments/:id" component={Department} />
-              <PrivateRoute path="/departments" component={Departments} />
-              <Route path="*" component={Error} />
-            </Switch>
-          </Suspense>
-        </Switch>
-      </Container>
-    </BrowserRouter>
+    <Container>
+      {children}
+    </Container>
   );
 }
