--- src/components/Department/Department.css
@@ -1,5 +1,0 @@
-.Department__infowrapper {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
