--- src/components/Departments/Departments.css
@@ -1,14 +1,0 @@
-.Departments {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.Departments__name {
-  text-align: left;
-}
-
-.Departments__btn--link {
-  color: #fff;
-  text-decoration: none;
-}
