+++ src/components/Login.jsx
@@ -1,0 +1,21 @@
+import React, { Component } from 'react';
+
+class Login extends Component {
+  constructor(props) {
+    super(props);
+    this.state = {
+
+    };
+  }
+
+  render() {
+    return (
+      <div>
+        <h2>Login Page</h2>
+        <a href="./departments">departments</a>
+      </div>
+    );
+  }
+}
+
+export default Login;
