--- src/components/Login/Login.css
@@ -1,4 +1,0 @@
-.Login__wrapper {
-  width: 100%;
-  max-width: 500px;
-}
