--- src/routes/PrivateRoute.jsx
@@ -1,23 +1,0 @@
-import React from 'react';
-import { Route, Redirect } from 'react-router-dom';
-
-export default function PrivateRoute({ component: Component, ...rest }) {
-  return (
-    (
-      <Route
-        {...rest}
-        render={(props) => (
-          localStorage.getItem(process.env.REACT_APP_TOKEN)
-            ? <Component {...props} />
-            : (
-              <Redirect to={{
-                pathname: '/',
-                state: { from: props.location },
-              }}
-              />
-            )
-        )}
-      />
-    )
-  );
-}
